restoring user and relating to project

+++ src/routes/api/stitch/+server.ts
@@ -35,24 +35,24 @@ export const DELETE: RequestHandler = async (event) => {
          id: stitchId
       },
 
-      // include: {
-      //    row: {
-      //       include: {
-      //          project: {
-      //             select: {
-      //                ownerId: true
-      //             }
-      //          }
-      //       }
-      //    }
-      // }
+      include: {
+         row: {
+            include: {
+               project: {
+                  select: {
+                     ownerId: true
+                  }
+               }
+            }
+         }
+      }
    })
    if(!stitch){
       return new Response(JSON.stringify('Stitch not found'), {status: 404})
    }
-   // if(stitch.row.project.ownerId !== event.locals.user?.id){
-   //    return new Response(JSON.stringify('Not your stitch'), { status: 402})
-   // }
+   if(stitch.row.project.ownerId !== event.locals.user?.id){
+      return new Response(JSON.stringify('Not your stitch'), { status: 402})
+   }
    await prisma.stitch.delete({
       where: {
          id: stitch.id
